Redesign state machine to be generic and add local controller.
This is a massive redesign of the internals as well as related changes to
the UI.  Most notably, the engine can now support different methods for
running MineCraft via generic, overridable State and StateMachine classes.
Included is the droplet model from OoD's previous incarnation and a simple
local server.  State timeouts are now configurable.  The database has been
rearranged and expanded.  Everything is in database now aside from a small
number of settings and droplet ssh and API keys.  The engine also supports
multiple MineCraft instances.

The simple server is standalone app manages a MineCraft server instance via a
basic TCP server.  The protocol is line based, with "start" starting the server
and "stop" stopping it.  Output from the app and stdout/stderr from MineCraft
is recorded to the given path, in a file named `ood_local.log`.  This is a
good way to test OoD locally.