Real state machine.
There's now a generic State and StateMachine class with implementations
for the simple server.  State timeouts are now configurable.  Database
rearranged and expanded.  Everything is in database now aside from droplet
ssh and API keys.

Web UI is currently broken; it needs to be expanded to understand multiple
OoD instances, as well as the new models.  The droplet stuff needs to be
completely refactored into the new state framework.