Mount local source in vagrant box.
Rather than cloning the ood source, mount the local source directory
inside the vagrant box.  Old code is commented out for now as this
would be useful in a proper production deployment scenario.

Note that this requires the vbguest plugin, installable via

    vagrant plugin install vagrant-vbguest

Also, upgrade Ansible dependency and map port 8000 from guest to
host for use in running the development server.